requirements.txt
asgiref==3.8.1
audioread==3.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.0.4
djangorestframework==3.15.1
filelock==3.13.3
fsspec==2024.3.1
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==24.0
platformdirs==4.2.0
pooch==1.8.1
pycparser==2.22
pydub==0.25.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.13.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.4.0
tiktoken==0.6.0
torch==2.2.2
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.11.0
urllib3==2.2.1
